requirements.txt
git+https://github.com/thisbejim/Pyrebase
discord.py==1.7.1
beautifulsoup4==4.9.3
requests==2.11.1
humanfriendly==9.1
numpy==1.19.0
numerize==0.12
PyYAML==5.3.1
pytz==2020.1
nest-asyncio==1.5.1
python-dotenv==0.17.1
